+++ FuelRod/Plutonium.cs
@@ -9,13 +9,13 @@ namespace Kerncentrale.FuelRod
     class Plutonium : FuelRod
     {
         private string name;
-        private int huidigeTemperatuur;
-        private int temperatuur;
-        private int optimaleTemperatuur;
+        private double huidigeTemperatuur;
+        private double temperatuur;
+        private double optimaleTemperatuur;
         private double graadPerLiter;
-        private int overhittingsTemperatuur;
-        private int onderLimietTemperatuur;
-        private int tempIncrease;
+        private double overhittingsTemperatuur;
+        private double onderLimietTemperatuur;
+        private double tempIncrease;
 
 
         public Plutonium()
@@ -42,24 +42,24 @@ namespace Kerncentrale.FuelRod
             }           
         }
 
-        public override void AfkoelenMetLitersWater(int waterInLiter){
+        public override void AfkoelenMetLitersWater(double waterInLiter){
             SetHuidigeTemperatuur(huidigeTemperatuur - (waterInLiter * graadPerLiter));
         }
 
         public override void SetGraadPerLiter(double value) => this.graadPerLiter = value;
 
-        public override void SetHuidigeTemperatuur(int value) => this.huidigeTemperatuur = value;
+        public override void SetHuidigeTemperatuur(double value) => this.huidigeTemperatuur = value;
 
         public override void SetName(string value) => this.name = value;
 
-        public override void SetOnderLimietTemperatuur(int value) => this.onderLimietTemperatuur = value;
+        public override void SetOnderLimietTemperatuur(double value) => this.onderLimietTemperatuur = value;
 
-        public override void SetOptimaleTemperatuur(int value) => this.optimaleTemperatuur = value;
+        public override void SetOptimaleTemperatuur(double value) => this.optimaleTemperatuur = value;
 
-        public override void SetOverhittingsTemperatuur(int value) => this.overhittingsTemperatuur = value;
+        public override void SetOverhittingsTemperatuur(double value) => this.overhittingsTemperatuur = value;
 
-        public override void SetTemperatuur(int value) => this.onderLimietTemperatuur = value;
+        public override void SetTemperatuur(double value) => this.onderLimietTemperatuur = value;
 
-        public override int SetTempIncrease(int value) => this.tempIncrease = value;
+        public override double SetTempIncrease(double value) => this.tempIncrease = value;
     }
 }
